Time: 19 ms (5.37%), Space: 33 MB (63.7%) - LeetHub

+++ 0703-kth-largest-element-in-a-stream/0703-kth-largest-element-in-a-stream.cpp
@@ -21,7 +21,7 @@ public:
         minHeap.push(val); 
         
         //  If the heap size exceeds k, remove the smallest element
-        if (minHeap.size() > (size_t)k) {
+        if (minHeap.size() > k) {
             minHeap.pop();
         }
